feat: implement jotai store in auth flow
