+++ src/contexts/index.ts
@@ -1,3 +1,2 @@
-export * from './AuthContext'
 export * from './NotificationContext'
 export * from './ColorSchemeContext'
