--- src/hooks/useAuth.ts
@@ -1,2 +1,0 @@
-// This was made to prevent require cycle in the app
-export { useAuthContext as useAuth } from '~contexts'
