+++ src/providers/index.ts
@@ -1,6 +1,5 @@
 export { NativeBaseProvider } from 'native-base'
 export { SafeAreaProvider } from 'react-native-safe-area-context'
-export * from './AuthProvider'
 export * from './NotificatedProvider'
 export * from './NotificationProvider'
 export * from './Providers'
