+++ src/store/auth/authState.ts
@@ -1,0 +1,3 @@
+import { atom } from 'jotai'
+
+export const isSignedInAtom = atom<boolean | null>(null)
