+++ src/store/store.ts
@@ -1,0 +1,3 @@
+import { createStore } from 'jotai'
+
+export const store = createStore()
