+++ src/utils/testUtils.tsx
@@ -5,7 +5,6 @@ import { PropsWithChildren, ReactElement } from 'react'
 import { I18nextProvider } from 'react-i18next'
 
 import i18n from '~i18n/i18nForTests'
-import { AuthProvider } from '~providers'
 import { ColorSchemeProvider } from '~providers/ColorSchemeProvider'
 
 type RenderOptions = Parameters<typeof render>[1]
@@ -16,15 +15,13 @@ const nbInitialWindowMetrics = {
 }
 
 const ProvidersWrapper: React.FC<PropsWithChildren> = ({ children }) => (
-  <AuthProvider>
-    <NativeBaseProvider initialWindowMetrics={nbInitialWindowMetrics}>
-      <NavigationContainer>
-        <ColorSchemeProvider>
-          <I18nextProvider i18n={i18n}>{children}</I18nextProvider>
-        </ColorSchemeProvider>
-      </NavigationContainer>
-    </NativeBaseProvider>
-  </AuthProvider>
+  <NativeBaseProvider initialWindowMetrics={nbInitialWindowMetrics}>
+    <NavigationContainer>
+      <ColorSchemeProvider>
+        <I18nextProvider i18n={i18n}>{children}</I18nextProvider>
+      </ColorSchemeProvider>
+    </NavigationContainer>
+  </NativeBaseProvider>
 )
 
 const customRender = (ui: ReactElement, options?: RenderOptions): RenderAPI => {
